requirements.txt
black==23.7.0
GitPython==3.1.32
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0.1
xmltodict==0.13.0
